5_ExplorationDataAnalysis/SimulatedIndustrialSensors/modules/SimulatedManufacturingSensors/requirements.txt
azure-iot-device~=2.7.0
pandas